Renamed .hpp to .h file.
