+++ applications/GeoMechanicsApplication/custom_elements/U_Pw_small_strain_interface_element.cpp
@@ -14,7 +14,7 @@
 // Application includes
 #include "custom_elements/U_Pw_small_strain_interface_element.hpp"
 
-#include <custom_utilities/stress_strain_utilities.hpp>
+#include <custom_utilities/stress_strain_utilities.h>
 
 namespace Kratos
 {
