+++ applications/GeoMechanicsApplication/custom_elements/updated_lagrangian_U_Pw_diff_order_axisymmetric_element.hpp
@@ -19,7 +19,7 @@
 // Application includes
 #include "custom_elements/updated_lagrangian_U_Pw_diff_order_element.hpp"
 #include "custom_utilities/element_utilities.hpp"
-#include "custom_utilities/stress_strain_utilities.hpp"
+#include "custom_utilities/stress_strain_utilities.h"
 #include "geo_mechanics_application_variables.h"
 
 namespace Kratos
