+++ applications/GeoMechanicsApplication/custom_utilities/stress_strain_utilities.h
