+++ applications/GeoMechanicsApplication/tests/cpp_tests/test_Hencky_strain.cpp
@@ -10,9 +10,9 @@
 //  Main authors:    Wijtze Pieter Kikstra
 //
 
-#include "testing/testing.h"
 #include "custom_utilities/math_utilities.hpp"
-#include "custom_utilities/stress_strain_utilities.hpp"
+#include "custom_utilities/stress_strain_utilities.h"
+#include "testing/testing.h"
 
 using namespace Kratos;
